third_party/blink/web_tests/fast/dom/shadow/apply-deep-in-document-scope-expected.txt

CONSOLE WARNING: /deep/ combinator is no longer supported in CSS dynamic profile. It is now effectively no-op, acting as if it were a descendant combinator. /deep/ combinator will be removed, and will be invalid at M65. You should remove it. See https://www.chromestatus.com/features/4964279606312960 for more details.
/deep/ as a descendant selector in document without shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inner).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

This text should be green
